Ensure all arguments are used
The argument processing using getopt was broken:
it used the first argument as the optstring, and
so discarded its value. This adds a new scenario
to catch the bug more specifically, and also fixes
the broken scenario from last time.

The other scenarios continued to pass only because,
with no target arguments, find searches everything
in the cwd.