Further refactoring to use new methods to be ensure files mtime is never changed after creation.
This commit had to deal with the mtimes of the directories and sub-dirs
created in the fixture setup. Although the files are given the required
mtimes, by default the mtime of the directories was the time that the
files were created. This could be after the "since" point, and so lead
to flickering tests. I've set the mtimes of the directories to match
the files, but it still leaves a potential problem if files are added
into subdirectories.