New feature: allow a child (dir or file) of a target to be excluded from the check.
Had to change the logic for the mtimes of the parent directories
of created files. Now, if a directory exists its mtime is left
unchanged. Previously, it would be updated to match the new file
if that file's mtime was later than the directory. It turned out
that logic wasn't actually used, and it was a hindrance when it
comes to setting up the file system state for the new acceptance
tests.